% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcsoap-package.R
\docType{package}
\name{calcsoap-package}
\alias{calcsoap}
\alias{calcsoap-package}
\title{calcsoap: What the Package Does in One 'Title Case' Line}
\description{
One paragraph description of what the package does as one or more full sentences.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/zachzIAM/calcsoap}
  \item Report bugs at \url{https://github.com/zachzIAM/calcsoap/issues}
}

}
\keyword{internal}
