% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcsoap-package.R
\docType{package}
\name{calcsoap-package}
\alias{calcsoap}
\alias{calcsoap-package}
\title{calcsoap: A sample SOAP client package with embedded C++ code and Rcpp}
\description{
Provides wrappers around the basic operations of a calculator SOAP service.
  The purpose of the package is to demonstrate how C++ source code generated via \code{gsoap}
  can be embedded in an R package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/zachzIAM/calcsoap}
  \item Report bugs at \url{https://github.com/zachzIAM/calcsoap/issues}
}

}
\keyword{internal}
