% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcsoap-package.R
\docType{package}
\name{calcsoap-package}
\alias{calcsoap}
\alias{calcsoap-package}
\title{calcsoap: What the Package Does in One 'Title Case' Line}
\description{
One paragraph description of what the package does as one or more full sentences.
}
\keyword{internal}
